actual_budget_importer/pyproject.toml
[project]
name = "hdfc-actual-budget-importer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "actualpy>=0.7.0",
    "pandas>=2.2.3",
    "pytest>=8.3.3",
    "streamlit>=1.40.1",
]
